Systematically Resolve the Post-condition of Moving Method
A method to be moved must be follow at least one of the following
conditions:
1) the method access a field of target type;
2) the method has a parameter of target type;

Therefore, given the aforementioned cases, the post condition is as
follows:
First, I search for the variable declaration resulting such a move, it
should be either a field declaration or a parameter declaration. Let me
call it root variable declaration.
Second, for every field access or method invocation in the moved method,
if its access object (e.g., a.m(), a is the accessing object of m) is
defined by root variable declaration, such an accessing object is
removed. Note that, the corresponding field access should also be
removed as well.
Third, I change the referers of this moved method. The cases are
distinguished into dimensions of field or parameter as well as outside
or inside the source class.